claimsense/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# ── Pydantic ──────────────────────────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Auth ──────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ── AI / LLM ─────────────────────────────────────────────────────────
google-generativeai>=0.8.0

# ── Notifications ─────────────────────────────────────────────────────
twilio>=9.0.0

# ── Encryption ────────────────────────────────────────────────────────
cryptography>=42.0.0

# ── FHIR ──────────────────────────────────────────────────────────────
fhir.resources>=7.1.0

# ── SSE ───────────────────────────────────────────────────────────────
sse-starlette>=2.0.0

# ── HTTP Client ───────────────────────────────────────────────────────
httpx>=0.27.0
